+++ src/main/java/blackjack/domain/card/Deck.java
@@ -1,10 +1,9 @@
-package blackjack.domain;
+package blackjack.domain.card;
 
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import java.util.stream.Stream;
-
 
 public class Deck {
