+++ src/main/java/blackjack/domain/card/ParticipantCards.java
@@ -1,8 +1,10 @@
-package blackjack.domain;
+package blackjack.domain.card;
 
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
+
+import blackjack.domain.card.Card;
 
 public class ParticipantCards {
