+++ src/main/java/blackjack/domain/participant/Dealer.java
@@ -1,4 +1,7 @@
-package blackjack.domain;
+package blackjack.domain.participant;
+
+import blackjack.domain.card.Card;
+import blackjack.domain.participant.Participant;
 
 public class Dealer extends Participant {
