+++ src/main/java/blackjack/domain/participant/Participant.java
@@ -1,8 +1,11 @@
-package blackjack.domain;
+package blackjack.domain.participant;
 
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
+
+import blackjack.domain.card.Card;
+import blackjack.domain.card.ParticipantCards;
 
 public abstract class Participant {
