+++ src/main/java/blackjack/domain/participant/Participants.java
@@ -1,8 +1,11 @@
-package blackjack.domain;
+package blackjack.domain.participant;
 
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
+
+import blackjack.domain.card.Card;
+import blackjack.domain.card.Deck;
 
 public class Participants {
