+++ src/test/java/blackjack/domain/ParticipantsTest.java
@@ -7,6 +7,9 @@ import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import blackjack.domain.participant.Participants;
+import blackjack.domain.participant.Player;
+
 public class ParticipantsTest {
 
     @Test
